+++ Media/Media.lua
@@ -1,0 +1,11 @@
+local LSM = LibStub("LibSharedMedia-3.0")
+
+if LSM == nil then
+  return
+end
+
+-- Audio
+LSM:Register("sound", "NT_Pet", [[Interface\AddOns\NoobTaco-GotOne\Media\Audio\NT_sd_pet.ogg]])
+LSM:Register("sound", "NT_Mount_Collection", [[Interface\AddOns\NoobTaco-GotOne\Media\Audio\NT_sd_mount_collection.ogg]])
+LSM:Register("sound", "NT_Toy_Collection", [[Interface\AddOns\NoobTaco-GotOne\Media\Audio\NT_sd_toy_collection.ogg]])
+LSM:Register("sound", "NT_Transmog", [[Interface\AddOns\NoobTaco-GotOne\Media\Audio\NT_sd_transmog.ogg]])